refactor: move errors pkg to internal
