+++ internal/errors/storage_errors.go
